// client-side/src/types/attendance.ts
export type Attendance = {
  
    _id:string;
    employeeId: string;
    clockInTime: string;
    formattedDate: string;
    clockOutTime:string;
    clockInDate: string;
    
    clockOutDate:string;
        underTime:string;
        lateTime: string;
        overtime: string;
    // email:string;
    // contactNo: string;
    // employeeRole: string;
    // employeeSalary: string;
    // address:string;
    // imagePath: string;
    
  };